requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain-openai
langchain
langchain-community
langchain-google-genai
langchain-chroma
firecrawl-py
langchainhub
wikipedia
chromadb
streamlit
gradio
pydantic
langchain-google-firestore
firestore
requests
langchain-huggingface
transformers
pypdf
pymongo